Refactor writers
