+++ mnelab/io/__init__.py
@@ -1,4 +1,5 @@
 from .readers import read_raw
+from .writers import write_raw, EXPORT_FORMATS
 
 
-__all__ = [read_raw]
+__all__ = [read_raw, write_raw, EXPORT_FORMATS]
