+++ mnelab/io/readers.py
@@ -4,19 +4,23 @@
 
 from pathlib import Path
 from functools import partial
+
 import mne
 
+from ..utils import have
+from .xdf import read_raw_xdf
 
-def _read_unknown(fname, **kwargs):
+
+def _read_unsupported(fname, **kwargs):
     ext = "".join(Path(fname).suffixes)
-    msg = f"Unknown file type ({ext})."
+    msg = f"Unsupported file type ({ext})."
     suggest = kwargs.get("suggest")
     if suggest is not None:
         msg += f" Try reading a {suggest} file instead."
     raise ValueError(msg)
 
 
-# TODO: what about multiple suffixes such as .xdf.gz?
+# supported read file formats
 supported = {".edf": mne.io.read_raw_edf,
              ".bdf": mne.io.read_raw_bdf,
              ".gdf": mne.io.read_raw_gdf,
@@ -28,9 +32,18 @@ supported = {".edf": mne.io.read_raw_edf,
              ".mff": mne.io.read_raw_egi,
              ".nxe": mne.io.read_raw_eximia,
              ".hdr": mne.io.read_raw_nirx}
-suggested = {".vmrk": partial(_read_unknown, suggest=".vhdr"),
-             ".eeg": partial(_read_unknown, suggest=".vhdr")}
-reader = {**supported, **suggested}
+
+if have["pyxdf"]:
+    supported.update({".xdf": read_raw_xdf,
+                      ".xdfz": read_raw_xdf,
+                      ".xdf.gz": read_raw_xdf})
+
+# known but unsupported file formats
+suggested = {".vmrk": partial(_read_unsupported, suggest=".vhdr"),
+             ".eeg": partial(_read_unsupported, suggest=".vhdr")}
+
+# all known file formats
+readers = {**supported, **suggested}
 
 
 def read_raw(fname, *args, **kwargs):
@@ -48,10 +61,13 @@ def read_raw(fname, *args, **kwargs):
 
     Notes
     -----
-
+    This function supports reading different file formats. It uses the readers
+    dict to dispatch the appropriate read function for a supported file type.
     """
     ext = "".join(Path(fname).suffixes)
-    if ext in reader:
-        return reader[ext](fname, *args, **kwargs)
+    if ext in readers:
+        return readers[ext](fname, *args, **kwargs)
     else:
         raise ValueError(f"Unknown file type {ext}.")
+        # here we could inspect the file signature to determine its type, which
+        # would allow us to read file independently of their extensions
