+++ mnelab/utils/__init__.py
@@ -3,10 +3,6 @@
 # License: BSD (3-clause)
 
 from .dependencies import have
-from .io import (IMPORT_FORMATS, EXPORT_FORMATS, split_fname, parse_xdf,
-                 parse_chunks, read_raw_xdf, get_xml, image_path)
-from .utils import has_locations
+from .utils import has_locations, image_path
 
-
-__all__ = [have, IMPORT_FORMATS, EXPORT_FORMATS, split_fname, parse_xdf,
-           parse_chunks, read_raw_xdf, get_xml, image_path]
+__all__ = [have, image_path]
